Rewrited unit tests logic: all pytest.fixtures are stored in a separate file && small bug fixes in test functions
